Round BMI value and fix POM
- correct the surefire and failsafe plugin version
- extract duplicate java version
- fix typo
- enable integration-test execution

+++ src/test/java/nl/goovaerts/api/BmiCalculationEndpointTest.java
@@ -6,12 +6,14 @@ import nl.goovaerts.data.BmiResponse;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 class BmiCalculationEndpointTest {
 
     @Test
     void shouldReturnBmiResponseForBmiRequest() {
         // Arrange
-        BmiResponse expected = new BmiResponse(22d);
+        BmiResponse expected = new BmiResponse(BigDecimal.valueOf(22d));
         BmiCalculationEndpoint endpoint = new BmiCalculationEndpoint(request -> expected);
 
         // Act
